1.04b
logDevice
WebThings Root to /things
Idx->Name
tracelevel
Eprom SizeWarning
Pages
memfix

+++ WAdapter/WAdapterMqtt.h
@@ -591,10 +591,10 @@ private:
 
 	// reads a byte into result
 	bool readByte(uint8_t *result) {
-		uint32_t previousMillis = millis();
+		unsigned long previousMillis = millis();
 		while (!_client->available()) {
 			yield();
-			uint32_t currentMillis = millis();
+			unsigned long currentMillis = millis();
 			if (currentMillis - previousMillis
 					>= ((int32_t) MQTT_SOCKET_TIMEOUT * 1000)) {
 				return false;
